requirements.txt
PyYAML==6.0.2
requests==2.32.4
packaging==24.1
tenacity==9.0.0
jsonschema==4.23.0

